backend/requirements.txt
fastapi
uvicorn
PyMuPDF
spacy
rapidfuzz
openai
python-dotenv
python-multipart
